Changed usage of configuration - now use ConfigurationField on the model
This means the JSONField is directly on the model rather than via a
foreignkey
