+++ core/configuration/fields.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, unicode_literals
 
 from django.contrib.postgres.fields import JSONField
 
-from configuration.models import (
+from .models import (
     get_configuration_default,
     get_configuration_choices,
     list_configuration_keys,
@@ -13,6 +13,11 @@ import six
 
 
 class ConfigurationDict(dict):
+    """Dictionary wrapper object for defined configuration keys.
+
+    Limits keys to defined configuration keys. Pulls values from instance with
+    fallback to default (if defined) values.
+    """
 
     def __contains__(self, key):
         """Return if the key is a defined configuration key.
@@ -20,8 +25,7 @@ class ConfigurationDict(dict):
         :param key: The key to lookup - should be formatted as 'appname.keyname'.
         :returns: True if the key is defined, else False.
         """
-        return key in self._dict or key in list_configuration_keys()
-
+        return key in list_configuration_keys()
 
     def __getitem__(self, key):
         """Return the value for the specified configuration key.
@@ -36,9 +40,9 @@ class ConfigurationDict(dict):
         if not isinstance(key, six.string_types):
             raise TypeError('Keys must be strings')
 
-        if key in self._dict:
-            return self._dict[key]
-        elif key in list_configuration_keys():
+        if dict.__contains__(self, key):  # get overridden value
+            return dict.__getitem__(self, key)
+        elif key in list_configuration_keys():  # get default value
             return get_configuration_default(key)
         else:
             raise KeyError('Configuration key "{}" is not defined'.format(key))
@@ -62,10 +66,10 @@ class ConfigurationDict(dict):
                 'Configuration value "{}" for key "{}" is not a valid choice. '
                 'Valid options include {}.'.format(value, key, choices))
 
-        self._dict[key] = value
+        dict.__setitem__(self, key, value)
 
     def __delitem__(self, key):
-        """Reset the instance value to default for the specified configuration key.
+        """Reset the instance value to default for the pecified configuration key.
 
         :param key: The key to reset the value for - should be formatted as 'appname.keyname'.
         :raises TypeError: If the configuration key is not a string.
@@ -77,13 +81,26 @@ class ConfigurationDict(dict):
         if key not in list_configuration_keys():
             raise KeyError('Configuration key "{}" is not defined'.format(key))
 
-        del self._dict[key]
+        dict.__delitem__(self, key)
 
 
 class ConfigurationField(JSONField):
+    """Wrapper field for JSONField that uses ConfigurationDict in python."""
+
+    def __init__(self, *args, **kwargs):
+        """Set default to ConfigurationDict."""
+        if 'default' not in kwargs:
+            kwargs['default'] = ConfigurationDict
+        super(ConfigurationField, self).__init__(*args, **kwargs)
 
     def from_db_value(self, value, expression, connection, context):
-        return ConfigurationDict(value)
+        """Convert value to ConfigurationDict."""
+        try:
+            value = ConfigurationDict(value)
+        except:
+            value = ConfigurationDict()
+        return value
 
     def get_prep_value(self, value):
-        return super(ConfigurationField, self).get_prep_value(value._dict)
+        """Defer to JSONField implementation."""
+        return super(ConfigurationField, self).get_prep_value(value)
