+++ core/tests/test_models.py
@@ -8,10 +8,7 @@ from django.utils import timezone
 
 from model_mommy import mommy
 
-from core.models import (Sample, Process, Project, Investigation, Milestone,
-                         AppConfigurationDefault, InstanceConfiguration,
-                         get_configuration_default, get_configuration_choices,
-                         list_configuration_keys)
+from core.models import Sample, Process, Project, Investigation, Milestone
 
 
 class TestSampleManager(TestCase):
@@ -651,266 +648,3 @@ class TestUser(TestCase):
         self.assertTrue(self.user.get_milestones('owner').exists())
         self.assertTrue(self.user.get_milestones('member').exists())
         self.assertTrue(self.user.get_milestones('viewer').exists())
-
-
-class TestAppConfiguration(TestCase):
-
-    def test_app_config_str(self):
-        choices = ['a', 'b', 'c']
-        config = mommy.make(AppConfigurationDefault, key='test.key',
-                                                     default_value='a',
-                                                     choices=choices)
-        self.assertEqual(str(config), 'test.key: a {}'.format(choices))
-
-    def test_get_configuration_default_invalid_type(self):
-        with self.assertRaises(AppConfigurationDefault.DoesNotExist):
-            get_configuration_default(1)
-
-    def test_get_configuration_default_invalid_empty_string(self):
-        with self.assertRaises(AppConfigurationDefault.DoesNotExist):
-            get_configuration_default('')
-
-    def test_get_configuration_default_invalid_key(self):
-        with self.assertRaises(AppConfigurationDefault.DoesNotExist):
-            get_configuration_default('app.key')
-
-    def test_get_configuration_default_none_defined(self):
-        config = mommy.make(AppConfigurationDefault, key='test.key')
-        value = get_configuration_default('test.key')
-        self.assertEqual(config.default_value, '')
-
-    def test_get_configuration_default_valid(self):
-        config = mommy.make(AppConfigurationDefault, key='test.key',
-                                                     default_value='value')
-        value = get_configuration_default('test.key')
-        self.assertEqual(config.default_value, value)
-
-    def test_get_configuration_choices_invalid_type(self):
-        with self.assertRaises(AppConfigurationDefault.DoesNotExist):
-            get_configuration_choices(1)
-
-    def test_get_configuration_choices_invalid_empty_string(self):
-        with self.assertRaises(AppConfigurationDefault.DoesNotExist):
-            get_configuration_choices('')
-
-    def test_get_configuration_choices_invalid_key(self):
-        with self.assertRaises(AppConfigurationDefault.DoesNotExist):
-            get_configuration_choices('app.key')
-
-    def test_get_configuration_choices_valid(self):
-        config = mommy.make(AppConfigurationDefault, key='test.key',
-                                                     default_value='a',
-                                                     choices=['a', 'b'])
-        choices = get_configuration_choices('test.key')
-        self.assertEqual(config.choices, choices)
-
-    def test_get_configuration_choices_none_defined(self):
-        config = mommy.make(AppConfigurationDefault, key='test.key')
-        choices = get_configuration_choices('test.key')
-        self.assertEqual(config.choices, [])
-
-    def test_list_configuration_keys_invalid_type(self):
-        with self.assertRaises(TypeError):
-            list_configuration_keys(1)
-
-    def test_list_configuration_keys_invalid_empty_string(self):
-        self.assertListEqual(list_configuration_keys(''), [])
-
-    def test_list_configuration_keys_no_app(self):
-        AppConfigurationDefault.objects.all().delete()
-        config = [
-            mommy.make(AppConfigurationDefault, key='test.key1'),
-            mommy.make(AppConfigurationDefault, key='test.key2'),
-            mommy.make(AppConfigurationDefault, key='test.key3'),
-            mommy.make(AppConfigurationDefault, key='alt.key1'),
-            mommy.make(AppConfigurationDefault, key='alt.key2'),
-        ]
-        keys = list_configuration_keys()
-        self.assertListEqual([c.key for c in config], keys)
-
-    def test_list_configuration_keys_with_app(self):
-        AppConfigurationDefault.objects.all().delete()
-        config = [
-            mommy.make(AppConfigurationDefault, key='test.key1'),
-            mommy.make(AppConfigurationDefault, key='test.key2'),
-            mommy.make(AppConfigurationDefault, key='test.key3'),
-            mommy.make(AppConfigurationDefault, key='alt.key1'),
-            mommy.make(AppConfigurationDefault, key='alt.key2'),
-        ]
-        keys = list_configuration_keys('test')
-        self.assertListEqual([c.key for c in config if 'test' in c.key], keys)
-
-    def test_instance_config_str(self):
-        config = mommy.make(InstanceConfiguration,
-                            configuration={'test.key1': 'value1',
-                                           'test.key2': 'value2'})
-        self.assertEqual(str(config.configuration), str(config))
-
-    def test_instance_config_contains_not_defined(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertFalse('test.key4' in instance)
-
-    def test_instance_config_contains_in_instance(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertTrue('test.key2' in instance)
-
-    def test_instance_config_contains_in_app(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertTrue('test.key3' in instance)
-
-    def test_instance_config_contains_in_both(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertTrue('test.key1' in instance)
-
-    def test_instance_config_getitem_not_defined(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        with self.assertRaises(KeyError):
-            instance['test.key4']
-
-    def test_instance_config_getitem_in_instance(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertEqual(instance['test.key2'], 'value2')
-
-    def test_instance_config_getitem_in_app(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertEqual(instance['test.key3'], 'default_value3')
-
-    def test_instance_config_getitem_in_both(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertEqual(instance['test.key1'], 'value1')
-
-    def test_instance_config_setitem_not_defined(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        with self.assertRaises(KeyError):
-            instance['test.key4'] = 'value4'
-
-    def test_instance_config_setitem_no_choices(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1')
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3')
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertEqual(instance['test.key3'], 'default_value3')
-        instance['test.key3'] = 'value3'
-        self.assertEqual(instance['test.key3'], 'value3')
-
-    def test_instance_config_setitem_invalid_choice(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1',
-                                choices=['default_value1', 'value1'])
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3',
-                                choices=['default_value3', 'value3'])
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        with self.assertRaises(ValueError):
-            instance['test.key3'] = 'invalid'
-
-    def test_instance_config_setitem_valid_choice(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1',
-                                choices=['default_value1', 'value1'])
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3',
-                                choices=['default_value3', 'value3'])
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1',
-                                             'test.key2': 'value2'})
-        self.assertEqual(instance['test.key3'], 'default_value3')
-        instance['test.key3'] = 'value3'
-        self.assertEqual(instance['test.key3'], 'value3')
-
-    def test_instance_config_reset_not_defined(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1',
-                                choices=['default_value1', 'value1'])
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3',
-                                choices=['default_value3', 'value3'])
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1'})
-        with self.assertRaises(KeyError):
-            instance.reset('test.key4')
-
-    def test_instance_config_reset_in_instance(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1',
-                                choices=['default_value1', 'value1'])
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3',
-                                choices=['default_value3', 'value3'])
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1'})
-        self.assertEqual(instance['test.key1'], 'value1')
-        instance.reset('test.key1')
-        self.assertEqual(instance['test.key1'], 'default_value1')
-
-    def test_instance_config_reset_in_app(self):
-        app_config = mommy.make(AppConfigurationDefault, key='test.key1',
-                                default_value='default_value1',
-                                choices=['default_value1', 'value1'])
-        app_config = mommy.make(AppConfigurationDefault, key='test.key3',
-                                default_value='default_value3',
-                                choices=['default_value3', 'value3'])
-        instance = mommy.make(InstanceConfiguration,
-                              configuration={'test.key1': 'value1'})
-        self.assertEqual(instance['test.key3'], 'default_value3')
-        instance.reset('test.key3')
-        self.assertEqual(instance['test.key3'], 'default_value3')
